Simple implementation of map files is underway. Testing now. We will make a resource to make/validate maps as we need them

+++ src/characters/WorldObject.java
@@ -20,6 +20,6 @@ public abstract class WorldObject {
 	public void setY(int y){ this.y = y; }
 	
 	//Get location
-	public double getX(){ return x; }
-	public double getY(){ return y; }
+	public int getX(){ return x; }
+	public int getY(){ return y; }
 }
